build: create a 'libc' library.

+++ libc/include/stdint.h
@@ -3,8 +3,8 @@
 #ifndef _STDINT_H_
 #define _STDINT_H_
 
-typedef signed short int16_t;
-typedef unsigned short uint16_t;
+typedef int int16_t;
+typedef unsigned int uint16_t;
 
 typedef signed char int8_t;
 typedef unsigned char uint8_t;
